feat: Add Alerts

+++ app/src/main/java/com/example/monitoringapp/data/network/request/DailyReportRequest.kt
@@ -2,9 +2,9 @@ package com.example.monitoringapp.data.network.request
 
 data class DailyReportRequest(
     var currentDate: String? = "",
-    var heartRate: Int? = 0,
-    var saturation: Int? = 0,
-    var temperature: Int? = 0
+    var heartRate: Double? = 0.0,
+    var saturation: Double? = 0.0,
+    var temperature: Double? = 0.0
 )
 
 data class DailyReportDateRequest(
